Ammended OAuthSecurity to better handle situation where insert is requested and secure parameter already exists.

+++ src/Material/OAuth/OAuthConfiguration.cs
@@ -5,6 +5,6 @@
         /// <summary>
         /// Default timeout for all oauth security parameters (state, nonce, etc)
         /// </summary>
-        public static int SecurityParameterTimeoutInMinutes { get; set; } = 2;
+        public static double SecurityParameterTimeoutInMinutes { get; set; } = 2;
     }
 }
